Establishing an algorithm that helps to identify NA values caused by sensor maintainance field-operations.
The algorithm uses the entry caused by a maintainance case "Protokolliert" as an anchor. From there he uses the RECORD numbers
to count +/- 1 and for safety measures also includes +/- 2 in forward and backward direction. Since the NA affected lines have been separated beforehand,
a gap in RECORD´s appear and therefore, marks the end/beginning of a specific sensor block. Other NA Values that are not maintainance-related and therefore have
no "Protokolliert" as an anchor will be excluded from this logic and marked for a more in-depth investigation. Since it is assumed that this errors have other causes.